refactor(frontmatter): Pull out as a dedicated mod
