+++ src/cargo/util/frontmatter.rs
@@ -1,0 +1,510 @@
+use crate::CargoResult;
+
+#[derive(Debug)]
+pub struct ScriptSource<'s> {
+    shebang: Option<&'s str>,
+    info: Option<&'s str>,
+    frontmatter: Option<&'s str>,
+    content: &'s str,
+}
+
+impl<'s> ScriptSource<'s> {
+    pub fn parse(input: &'s str) -> CargoResult<Self> {
+        let mut source = Self {
+            shebang: None,
+            info: None,
+            frontmatter: None,
+            content: input,
+        };
+
+        if let Some(shebang_end) = strip_shebang(source.content) {
+            let (shebang, content) = source.content.split_at(shebang_end);
+            source.shebang = Some(shebang);
+            source.content = content;
+        }
+
+        let mut rest = source.content;
+
+        // Whitespace may precede a frontmatter but must end with a newline
+        let trimmed = rest.trim_start_matches(is_whitespace);
+        if trimmed.len() != rest.len() {
+            let trimmed_len = rest.len() - trimmed.len();
+            let last_trimmed_index = trimmed_len - 1;
+            if rest.as_bytes()[last_trimmed_index] != b'\n' {
+                // either not a frontmatter or invalid opening
+                return Ok(source);
+            }
+        }
+        rest = trimmed;
+
+        // Opens with a line that starts with 3 or more `-` followed by an optional identifier
+        const FENCE_CHAR: char = '-';
+        let fence_length = rest
+            .char_indices()
+            .find_map(|(i, c)| (c != FENCE_CHAR).then_some(i))
+            .unwrap_or(rest.len());
+        match fence_length {
+            0 => {
+                return Ok(source);
+            }
+            1 | 2 => {
+                // either not a frontmatter or invalid frontmatter opening
+                anyhow::bail!(
+                    "found {fence_length} `{FENCE_CHAR}` in rust frontmatter, expected at least 3"
+                )
+            }
+            _ => {}
+        }
+        let (fence_pattern, rest) = rest.split_at(fence_length);
+        let Some(info_end_index) = rest.find('\n') else {
+            anyhow::bail!("no closing `{fence_pattern}` found for frontmatter");
+        };
+        let (info, rest) = rest.split_at(info_end_index);
+        let info = info.trim_matches(is_whitespace);
+        if !info.is_empty() {
+            source.info = Some(info);
+        }
+
+        // Ends with a line that starts with a matching number of `-` only followed by whitespace
+        let nl_fence_pattern = format!("\n{fence_pattern}");
+        let Some(frontmatter_nl) = rest.find(&nl_fence_pattern) else {
+            anyhow::bail!("no closing `{fence_pattern}` found for frontmatter");
+        };
+        let frontmatter = &rest[..frontmatter_nl + 1];
+        let frontmatter = frontmatter
+            .strip_prefix('\n')
+            .expect("earlier `found` + `split_at` left us here");
+        source.frontmatter = Some(frontmatter);
+        let rest = &rest[frontmatter_nl + nl_fence_pattern.len()..];
+
+        let (after_closing_fence, rest) = rest.split_once("\n").unwrap_or((rest, ""));
+        let after_closing_fence = after_closing_fence.trim_matches(is_whitespace);
+        if !after_closing_fence.is_empty() {
+            // extra characters beyond the original fence pattern, even if they are extra `-`
+            anyhow::bail!("trailing characters found after frontmatter close");
+        }
+
+        let frontmatter_len = input.len() - rest.len();
+        source.content = &input[frontmatter_len..];
+
+        let repeat = Self::parse(source.content)?;
+        if repeat.frontmatter.is_some() {
+            anyhow::bail!("only one frontmatter is supported");
+        }
+
+        Ok(source)
+    }
+
+    pub fn shebang(&self) -> Option<&'s str> {
+        self.shebang
+    }
+
+    pub fn info(&self) -> Option<&'s str> {
+        self.info
+    }
+
+    pub fn frontmatter(&self) -> Option<&'s str> {
+        self.frontmatter
+    }
+
+    pub fn content(&self) -> &'s str {
+        self.content
+    }
+}
+
+fn strip_shebang(input: &str) -> Option<usize> {
+    // See rust-lang/rust's compiler/rustc_lexer/src/lib.rs's `strip_shebang`
+    // Shebang must start with `#!` literally, without any preceding whitespace.
+    // For simplicity we consider any line starting with `#!` a shebang,
+    // regardless of restrictions put on shebangs by specific platforms.
+    if let Some(rest) = input.strip_prefix("#!") {
+        // Ok, this is a shebang but if the next non-whitespace token is `[`,
+        // then it may be valid Rust code, so consider it Rust code.
+        //
+        // NOTE: rustc considers line and block comments to be whitespace but to avoid
+        // any more awareness of Rust grammar, we are excluding it.
+        if !rest.trim_start().starts_with('[') {
+            // No other choice than to consider this a shebang.
+            let newline_end = input.find('\n').map(|pos| pos + 1).unwrap_or(input.len());
+            return Some(newline_end);
+        }
+    }
+    None
+}
+
+/// True if `c` is considered a whitespace according to Rust language definition.
+/// See [Rust language reference](https://doc.rust-lang.org/reference/whitespace.html)
+/// for definitions of these classes.
+///
+/// See rust-lang/rust's compiler/rustc_lexer/src/lib.rs `is_whitespace`
+fn is_whitespace(c: char) -> bool {
+    // This is Pattern_White_Space.
+    //
+    // Note that this set is stable (ie, it doesn't change with different
+    // Unicode versions), so it's ok to just hard-code the values.
+
+    matches!(
+        c,
+        // Usual ASCII suspects
+        '\u{0009}'   // \t
+        | '\u{000A}' // \n
+        | '\u{000B}' // vertical tab
+        | '\u{000C}' // form feed
+        | '\u{000D}' // \r
+        | '\u{0020}' // space
+
+        // NEXT LINE from latin1
+        | '\u{0085}'
+
+        // Bidi markers
+        | '\u{200E}' // LEFT-TO-RIGHT MARK
+        | '\u{200F}' // RIGHT-TO-LEFT MARK
+
+        // Dedicated whitespace characters from Unicode
+        | '\u{2028}' // LINE SEPARATOR
+        | '\u{2029}' // PARAGRAPH SEPARATOR
+    )
+}
+
+#[cfg(test)]
+mod test {
+    use snapbox::assert_data_eq;
+    use snapbox::prelude::*;
+    use snapbox::str;
+
+    use super::*;
+
+    #[track_caller]
+    fn assert_source(source: &str, expected: impl IntoData) {
+        use std::fmt::Write as _;
+
+        let actual = match ScriptSource::parse(source) {
+            Ok(actual) => actual,
+            Err(err) => panic!("unexpected err: {err}"),
+        };
+
+        let mut rendered = String::new();
+        write_optional_field(&mut rendered, "shebang", actual.shebang());
+        write_optional_field(&mut rendered, "info", actual.info());
+        write_optional_field(&mut rendered, "frontmatter", actual.frontmatter());
+        writeln!(&mut rendered, "content: {:?}", actual.content()).unwrap();
+        assert_data_eq!(rendered, expected.raw());
+    }
+
+    fn write_optional_field(writer: &mut dyn std::fmt::Write, field: &str, value: Option<&str>) {
+        if let Some(value) = value {
+            writeln!(writer, "{field}: {value:?}").unwrap();
+        } else {
+            writeln!(writer, "{field}: None").unwrap();
+        }
+    }
+
+    #[track_caller]
+    fn assert_err(
+        result: Result<impl std::fmt::Debug, impl std::fmt::Display>,
+        err: impl IntoData,
+    ) {
+        match result {
+            Ok(d) => panic!("unexpected Ok({d:#?})"),
+            Err(actual) => snapbox::assert_data_eq!(actual.to_string(), err.raw()),
+        }
+    }
+
+    #[test]
+    fn split_default() {
+        assert_source(
+            r#"fn main() {}
+"#,
+            str![[r#"
+shebang: None
+info: None
+frontmatter: None
+content: "fn main() {}\n"
+
+"#]],
+        );
+    }
+
+    #[test]
+    fn split_dependencies() {
+        assert_source(
+            r#"---
+[dependencies]
+time="0.1.25"
+---
+fn main() {}
+"#,
+            str![[r#"
+shebang: None
+info: None
+frontmatter: "[dependencies]\ntime=\"0.1.25\"\n"
+content: "fn main() {}\n"
+
+"#]],
+        );
+    }
+
+    #[test]
+    fn split_infostring() {
+        assert_source(
+            r#"---cargo
+[dependencies]
+time="0.1.25"
+---
+fn main() {}
+"#,
+            str![[r#"
+shebang: None
+info: "cargo"
+frontmatter: "[dependencies]\ntime=\"0.1.25\"\n"
+content: "fn main() {}\n"
+
+"#]],
+        );
+    }
+
+    #[test]
+    fn split_infostring_whitespace() {
+        assert_source(
+            r#"--- cargo 
+[dependencies]
+time="0.1.25"
+---
+fn main() {}
+"#,
+            str![[r#"
+shebang: None
+info: "cargo"
+frontmatter: "[dependencies]\ntime=\"0.1.25\"\n"
+content: "fn main() {}\n"
+
+"#]],
+        );
+    }
+
+    #[test]
+    fn split_shebang() {
+        assert_source(
+            r#"#!/usr/bin/env cargo
+---
+[dependencies]
+time="0.1.25"
+---
+fn main() {}
+"#,
+            str![[r##"
+shebang: "#!/usr/bin/env cargo\n"
+info: None
+frontmatter: "[dependencies]\ntime=\"0.1.25\"\n"
+content: "fn main() {}\n"
+
+"##]],
+        );
+    }
+
+    #[test]
+    fn split_crlf() {
+        assert_source(
+            "#!/usr/bin/env cargo\r\n---\r\n[dependencies]\r\ntime=\"0.1.25\"\r\n---\r\nfn main() {}",
+            str![[r##"
+shebang: "#!/usr/bin/env cargo\r\n"
+info: None
+frontmatter: "[dependencies]\r\ntime=\"0.1.25\"\r\n"
+content: "fn main() {}"
+
+"##]],
+        );
+    }
+
+    #[test]
+    fn split_leading_newlines() {
+        assert_source(
+            r#"#!/usr/bin/env cargo
+    
+
+
+---
+[dependencies]
+time="0.1.25"
+---
+
+
+fn main() {}
+"#,
+            str![[r##"
+shebang: "#!/usr/bin/env cargo\n"
+info: None
+frontmatter: "[dependencies]\ntime=\"0.1.25\"\n"
+content: "\n\nfn main() {}\n"
+
+"##]],
+        );
+    }
+
+    #[test]
+    fn split_attribute() {
+        assert_source(
+            r#"#[allow(dead_code)]
+---
+[dependencies]
+time="0.1.25"
+---
+fn main() {}
+"#,
+            str![[r##"
+shebang: None
+info: None
+frontmatter: None
+content: "#[allow(dead_code)]\n---\n[dependencies]\ntime=\"0.1.25\"\n---\nfn main() {}\n"
+
+"##]],
+        );
+    }
+
+    #[test]
+    fn split_extra_dash() {
+        assert_source(
+            r#"#!/usr/bin/env cargo
+----------
+[dependencies]
+time="0.1.25"
+----------
+
+fn main() {}"#,
+            str![[r##"
+shebang: "#!/usr/bin/env cargo\n"
+info: None
+frontmatter: "[dependencies]\ntime=\"0.1.25\"\n"
+content: "\nfn main() {}"
+
+"##]],
+        );
+    }
+
+    #[test]
+    fn split_too_few_dashes() {
+        assert_err(
+            ScriptSource::parse(
+                r#"#!/usr/bin/env cargo
+--
+[dependencies]
+time="0.1.25"
+--
+fn main() {}
+"#,
+            ),
+            str!["found 2 `-` in rust frontmatter, expected at least 3"],
+        );
+    }
+
+    #[test]
+    fn split_indent() {
+        assert_source(
+            r#"#!/usr/bin/env cargo
+    ---
+    [dependencies]
+    time="0.1.25"
+    ----
+
+fn main() {}
+"#,
+            str![[r##"
+shebang: "#!/usr/bin/env cargo\n"
+info: None
+frontmatter: None
+content: "    ---\n    [dependencies]\n    time=\"0.1.25\"\n    ----\n\nfn main() {}\n"
+
+"##]],
+        );
+    }
+
+    #[test]
+    fn split_escaped() {
+        assert_source(
+            r#"#!/usr/bin/env cargo
+-----
+---
+---
+-----
+
+fn main() {}
+"#,
+            str![[r##"
+shebang: "#!/usr/bin/env cargo\n"
+info: None
+frontmatter: "---\n---\n"
+content: "\nfn main() {}\n"
+
+"##]],
+        );
+    }
+
+    #[test]
+    fn split_invalid_escaped() {
+        assert_err(
+            ScriptSource::parse(
+                r#"#!/usr/bin/env cargo
+---
+-----
+-----
+---
+
+fn main() {}
+"#,
+            ),
+            str!["trailing characters found after frontmatter close"],
+        );
+    }
+
+    #[test]
+    fn split_dashes_in_body() {
+        assert_source(
+            r#"#!/usr/bin/env cargo
+---
+Hello---
+World
+---
+
+fn main() {}
+"#,
+            str![[r##"
+shebang: "#!/usr/bin/env cargo\n"
+info: None
+frontmatter: "Hello---\nWorld\n"
+content: "\nfn main() {}\n"
+
+"##]],
+        );
+    }
+
+    #[test]
+    fn split_mismatched_dashes() {
+        assert_err(
+            ScriptSource::parse(
+                r#"#!/usr/bin/env cargo
+---
+[dependencies]
+time="0.1.25"
+----
+fn main() {}
+"#,
+            ),
+            str!["trailing characters found after frontmatter close"],
+        );
+    }
+
+    #[test]
+    fn split_missing_close() {
+        assert_err(
+            ScriptSource::parse(
+                r#"#!/usr/bin/env cargo
+---
+[dependencies]
+time="0.1.25"
+fn main() {}
+"#,
+            ),
+            str!["no closing `---` found for frontmatter"],
+        );
+    }
+}
