+++ src/cargo/util/mod.rs
@@ -41,6 +41,7 @@ pub mod diagnostic_server;
 pub mod edit_distance;
 pub mod errors;
 mod flock;
+pub mod frontmatter;
 pub mod graph;
 mod hasher;
 pub mod hex;
